phần quyền

+++ mvc/controllers/Author.php
@@ -3,10 +3,16 @@ class Author extends Controller{
     
     function __construct(){
         $this->author = $this->model("AuthorModel");
-        
+        $this->UserRole = $this->model("UserRoleModel");
+        $this->RolePermission = $this->model("RolePermissionModel");                        
+        if($this->UserRole->checkRole("staff.product")!=1 && $this->UserRole->checkRole("admin")!=1)
+        {
+            $this->page500();
+            exit();
+        }                                   
     }
 
-    function index(){
+    function index(){        
         $this->view("admin/layout",array(
 			"Page" => "author"       
 		));        
@@ -20,7 +26,11 @@ class Author extends Controller{
         echo $list;
     }
     function add(){
-        
+        if($this->UserRole->checkRole("admin")!=1 && $this->UserRole->checkPermission($_SESSION['username'],"staff.product","add")!=1)        
+        {
+            echo 2;
+            return;
+        }
         if(isset($_POST['txtName']) && $_POST['txtDetail']){
 
             $name = $_POST['txtName'];
@@ -32,12 +42,15 @@ class Author extends Controller{
                 return;
             }
         }
-        echo 0;
-
+        echo 0;                       
     }
 
     function update($id){
-        
+        if($this->UserRole->checkRole("admin")!=1 && $this->UserRole->checkPermission($_SESSION['username'],"staff.product","update")!=1)        
+        {
+            echo 2;
+            return;
+        }
         if(isset($_POST['txtName']) && $_POST['txtDetail']){
             
             $name = $_POST['txtName'];
@@ -53,6 +66,11 @@ class Author extends Controller{
         echo 0;
     }
     function delete($id){
+        if($this->UserRole->checkRole("admin")!=1 && $this->UserRole->checkPermission($_SESSION['username'],"staff.product","delete")!=1)        
+        {
+            echo 2;
+            return;
+        }
         if(isset($_POST['checkDelete'])){
             
             if($this->author->delete($id)==1){
@@ -61,9 +79,14 @@ class Author extends Controller{
             }
         }
         echo 0;
-    }
+    }      
     function pages() {
         $this->view("pages/404");
     }
+    function page500() {
+        $this->view("layout2",array(
+            "Page" => "500"
+        ));
+    }        
 }
 ?>
